sales_data_agent_expose/requirements.txt
google-genai
google-adk
toolbox-core
a2a-sdk[http-server]>=0.3.0